% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coiCheckR-package.R
\docType{package}
\name{coiCheckR-package}
\alias{coiCheckR}
\alias{coiCheckR-package}
\title{coiCheckR: Screen Peer Reviewer Candidates for Conflicts of Interest}
\description{
See the README and \code{vignette("coiCheckR")} for a walkthrough. Start
with \code{\link[=checkCoi]{checkCoi()}} for a single candidate or \code{\link[=checkCoiBatch]{checkCoiBatch()}} for
a shortlist.
}
\section{Rate limits}{

PubMed E-utilities allow 3 requests/sec without a key, 10/sec with
one. If you're screening more than a handful of names, get a free key
at \url{https://www.ncbi.nlm.nih.gov/account/settings/} and call
\code{\link[=setNcbiKey]{setNcbiKey()}} once per session.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jravilab/coiCheckR}
  \item Report bugs at \url{https://github.com/jravilab/coiCheckR/issues}
}

}
\author{
\strong{Maintainer}: Janani Ravi \email{janani.ravi@cuanschutz.edu}

Authors:
\itemize{
  \item Janani Ravi \email{janani.ravi@cuanschutz.edu}
}

}
\keyword{internal}
